Day-3/Langfuse/requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
chromadb
python-dotenv